+++ src/Facades/Vandar.php
@@ -20,6 +20,7 @@ use Zarbinco\LaravelVandar\LaravelVandar;
  * @method static \Zarbinco\LaravelVandar\Resources\CustomerResource customers()
  * @method static \Zarbinco\LaravelVandar\Resources\CardResource cards()
  * @method static \Zarbinco\LaravelVandar\Resources\IbanResource ibans()
+ * @method static \Zarbinco\LaravelVandar\Resources\InquiryResource inquiries()
  * @method static \Zarbinco\LaravelVandar\DTO\VandarResponse response(array $payload = [], int $status = 200, array $headers = [])
  *
  * @see LaravelVandar